drm/radeon: try a hard reset is resume failed after a soft reset
In case atom_asic_init fails with something like:
  *ERROR* atombios stuck in loop for more than 5secs aborting
  *ERROR* atombios stuck executing C483 (len 254, WS 0, PS 4) @ 0xC4AD
  *ERROR* atombios stuck executing BC59 (len 74, WS 0, PS 8) @ 0xBC8E
after a soft reset then it is worth a try to do a gpu hard reset.

Indeed when radeon_resume fails the vm_manager is off and user space
will be able to use the gpu anymore. Also it might lead to a machine
freeze. And un most of the case I have seen doing a hard reset allow
to recover the gpu.

Signed-off-by: Julien Isorce <jisorce@oblong.com>